Create “SpatialElement” class; move element classes from lib/ to element/; add property observation to elements.
First up, I’ve created a SpatialElement class that has the position, rotation, scale arguments, and Element can be used as a base class for elements that don’t actually have a spatial representation, such as Audio.

Element has some helper methods for defining Vector, Euler, Scalar, and String properties.

Each of the properties created with these helpers can have changes monitored with addPropertyChangeObserver() on Element. The core DOM classes now have better support for monitoring changes. This support is synchronous, rather than using the asychronous Object.observe, mainly because it makes it easier to reason about what's going to happen, and to test.

addPropertyChangeObserver() will look for deep-changes, so that el.position.x = val will trigger the position change.

Arguably it would be better to do this with Object.observe; regardless, addPropertyChangeObserver() will be a useful API method to have as it only calls in the case of a specific property change. If this code is refactored to use Object.observe in the future, the addPropertyChangeObserver() method on scene dom elements could be retained.