Refactor server to rely on SceneDOM for XML parsing.
Previously, the server was directly accessing the Scene object of SceneDOM rather than relying on the createDocumentFromXML() method. This ensures that SceneDOM preserves its encapsulation a little better.

This patch also refactors responsibility for setting up reflectors into the websocket server class, which is now initialised by passing it sets of SceneDOM documents and associated URLs.